# docs/upgrade/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReVector Core Engine",
    "description": "True editable geometry. No raster wrappers. Deploy behind JerseyOS authentication.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/revector/optimize": {
      "post": {
        "tags": [
          "processing jobs"
        ],
        "summary": "Optimize",
        "operationId": "optimize_api_revector_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/revector/prepare": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Prepare",
        "operationId": "prepare_api_revector_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/production": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Production",
        "operationId": "production_api_revector_production_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/recover-part": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Recover",
        "operationId": "recover_api_revector_recover_part_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/ai-missing": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Missing",
        "operationId": "missing_api_revector_ai_missing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/review/confirm": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Review",
        "operationId": "review_api_revector_review_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/slots/update": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Update Slot",
        "operationId": "update_slot_api_revector_slots_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/events": {
      "get": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Events",
        "operationId": "events_api_revector_projects__project_id__events_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/errors": {
      "get": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Errors",
        "operationId": "errors_api_revector_projects__project_id__errors_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/assistant/explain": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Explain",
        "operationId": "explain_api_revector_assistant_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/assistant/feedback": {
      "post": {
        "tags": [
          "AI-assisted production and error recovery"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_revector_assistant_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssistantRequest": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "error_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Id"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "question": {
            "type": "string",
            "maxLength": 1000,
            "title": "Question",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AssistantRequest"
      },
...
      "ErrorEvent": {
        "properties": {
          "error_id": {
            "type": "string",
            "title": "Error Id"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "part_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "technical_summary": {
            "type": "string",
            "title": "Technical Summary"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "recoverable": {
            "type": "boolean",
            "title": "Recoverable"
          },
          "suggested_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Actions"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "error_id",
          "error_code",
          "code",
          "category",
          "phase",
          "message",
          "technical_summary",
          "retryable",
          "recoverable",
          "timestamp"
        ],
        "title": "ErrorEvent"
      },
...
      "FeedbackRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "error_id": {
            "type": "string",
            "title": "Error Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Outcome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "error_id",
          "action",
          "outcome"
        ],
        "title": "FeedbackRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "job_state": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Job State"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEvent"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "process_event": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Event"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "project_id",
          "stage",
          "status",
          "job_state",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse"
      },
...
      "MissingRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "slot": {
            "type": "string",
            "enum": [
              "LEFT_SLEEVE",
              "RIGHT_SLEEVE",
              "FRONT_BODY",
              "BACK_BODY",
              "FRONT_COLLAR",
              "BACK_COLLAR",
              "TOP_TRIM",
              "BOTTOM_TRIM"
            ],
            "title": "Slot"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "slot"
        ],
        "title": "MissingRequest"
      },
...
      "RecoveryRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "fallback_trace": {
            "type": "boolean",
            "title": "Fallback Trace",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "RecoveryRequest"
      },
      "ReviewRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "decisions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SlotDecision"
            },
            "propertyNames": {
              "enum": [
                "LEFT_SLEEVE",
                "RIGHT_SLEEVE",
                "FRONT_BODY",
                "BACK_BODY",
                "FRONT_COLLAR",
                "BACK_COLLAR",
                "TOP_TRIM",
                "BOTTOM_TRIM"
              ]
            },
            "type": "object",
            "title": "Decisions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "decisions"
        ],
        "title": "ReviewRequest"
      },
...
      "SlotDecision": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "blank"
            ],
            "title": "Status"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SlotDecision"
      },
      "SlotRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "slot": {
            "type": "string",
            "enum": [
              "LEFT_SLEEVE",
              "RIGHT_SLEEVE",
              "FRONT_BODY",
              "BACK_BODY",
              "FRONT_COLLAR",
              "BACK_COLLAR",
              "TOP_TRIM",
              "BOTTOM_TRIM"
            ],
            "title": "Slot"
          },
          "status": {
            "type": "string",
            "enum": [
              "blank",
              "missing"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "slot",
          "status"
        ],
        "title": "SlotRequest"
      },
      "StageRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "StageRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}